jjc: Fix incoming messages having a wildcard as target param
Server-wide notifications are usually send using a wildcard as target,
matching all users on the network.  Because jjc uses that target
parameter, to determine if a message is a channel or private message, it
would confuse a wildcard with the target being a channel instead of a
private message (normal private messages have our nick as the target).
That causes the creation of weird channels like "$$*" instead of a channel
with the actual source of the message.  This fixes that issue by checking
if the target channel exists and use the source as target if it does not.