add support for UCSPI, another way to support TLS or other protocols
UCSPI[1] is a convention for splitting the network handling code from
the protocol code: the protocol program (jjc) expects an open socket
to the network to be already open at file descriptor 6 (for reading
messages from) and 7 (to write messages to) when it starts.

Tools like s6-networking[2] can then bring TLS support to jj through
bearssl.org or libressl.org (with plan to support client certificates).

After this patch, jj can be started this way:
	s6-tlsclient chat.freenode.net 6697 jjd

It will detect that it is running under ucspi thanks to the PROTO
environment variable specified by UCSPI, so there is no need to
configure it, yet it can still use a regular TCP connexion for use
with stunnel[3] or inetd+openssl's s_client[4] as it currently can.

[1]: https://cr.yp.to/proto/ucspi.txt
[2]: https://skarnet.org/software/s6-networking/
[3]: https://www.stunnel.org/
[4]: http://tools.suckless.org/ii/usage/