flush the output buffer to force the last line to be printed
Some awk implementation keep the last line in the buffer, and do not print it
until a '\n' is printed.  Given that jjp uses printf("\n%s") rather than
printf("%s\n"), the awk implementation that are line-buffered will not print
the last line without that.

Note that POSIX awk(1p) does not have this function, and system("") could
be used as a workaround instead, which would have the side effect to spawn
a shell doing nothing for every line printed.