keep nawk compatibility
It looks like nawk does not support deleting anything that is not an array.

While delete keyword is used with an variable, the awk syntax parser set the
variable type to "array", and produce this error:

	/usr/bin/awk: can't assign to header; it's an array name.
	 input record number 2, file
	 source line number 58

By loosing a bit of rigor (not deleting an global variable), we earn a bit
of portability.