jjc: Add better way of only running external command if it can be found in $PATH
In `ext_command()`, we removed the code which first checked if the command can
be found in $PATH, before actually invoking it via `system()`. The result was
that `jjd(1)` would spam stderr on every call of `ext_command()` (given the
command was not in $PATH). This change handles non-existent commands more
gracefully while also saving one `system()` call.